refactor(plot): split image and console renderers by IR
Each renderer mixed two independent concerns: Figure charts and the
ActivationDiagram node-link graph. Split them into image/ and console/
folders with a shared mod.rs (config + color helpers) and figure.rs /
diagram.rs per IR, mirroring the scene/activations split on the pure
side. No behavior change.
