+++ core/src/plot/image/diagram.rs
@@ -1,93 +1,15 @@
-//! Rasterizing a [`Figure`] to an RGB image with `plotters`, and an
-//! [`ActivationDiagram`] to a horizontal node-link graph.
+//! Rasterizing an [`ActivationDiagram`] to a horizontal node-link graph: one
+//! column of neurons per layer, edges colored by weight sign, values annotated
+//! at the ends, and a legend along the bottom.
 
+use super::{ImageConfig, RasterImage, rgb};
 use crate::plot::activations::ActivationDiagram;
-use crate::plot::scene::{Color as SceneColor, Figure, Panel, Series};
+use crate::plot::scene::Color as SceneColor;
 use plotters::backend::BitMapBackend;
-use plotters::chart::{ChartBuilder, ChartContext};
 use plotters::coord::Shift;
-use plotters::coord::cartesian::Cartesian2d;
-use plotters::coord::types::RangedCoordf32;
 use plotters::prelude::*;
 use plotters::style::text_anchor::{HPos, Pos, VPos};
 use std::error::Error;
-
-/// Rendering options for rasterizing a [`Figure`].
-pub struct ImageConfig<'a> {
-    /// Width of the image in pixels.
-    pub width: u32,
-    /// Height of the image in pixels.
-    pub height: u32,
-    /// Font family for titles, labels and legends, e.g. "sans-serif".
-    pub font_style: &'a str,
-    /// Font size for text elements.
-    pub font_size: u32,
-    /// Size of the area reserved for axis labels and legends.
-    pub area_size: u32,
-}
-
-impl Default for ImageConfig<'_> {
-    fn default() -> Self {
-        Self {
-            width: 1200,
-            height: 900,
-            font_style: "sans-serif",
-            font_size: 20,
-            area_size: 40,
-        }
-    }
-}
-
-impl ImageConfig<'_> {
-    /// An image configuration of the given size, with default fonts and spacing.
-    pub fn new(width: u32, height: u32) -> Self {
-        Self {
-            width,
-            height,
-            ..Default::default()
-        }
-    }
-}
-
-/// A rendered RGB image: a row-major buffer of `width × height` 8-bit RGB triples.
-pub struct RasterImage {
-    /// Row-major RGB pixel data, three bytes per pixel.
-    pub bytes: Vec<u8>,
-    /// Image width in pixels.
-    pub width: u32,
-    /// Image height in pixels.
-    pub height: u32,
-}
-
-impl Figure {
-    /// Rasterizes the figure to a [`RasterImage`], stacking its panels vertically in order.
-    pub fn to_image(&self, cfg: &ImageConfig) -> Result<RasterImage, Box<dyn Error>> {
-        let mut bytes = vec![255u8; (cfg.width * cfg.height * 3) as usize];
-        {
-            let root =
-                BitMapBackend::with_buffer(&mut bytes, (cfg.width, cfg.height)).into_drawing_area();
-            root.fill(&WHITE)?;
-
-            let regions = root.split_evenly((self.panels.len(), 1));
-            for (panel, region) in self.panels.iter().zip(regions) {
-                draw_panel(&region, panel, cfg)?;
-            }
-
-            root.present()?;
-        }
-
-        Ok(RasterImage {
-            bytes,
-            width: cfg.width,
-            height: cfg.height,
-        })
-    }
-}
-
-/// The plotters color for a scene color.
-fn rgb(color: SceneColor) -> RGBColor {
-    RGBColor(color.red, color.green, color.blue)
-}
 
 /// Blank pixels reserved around the node field on every side.
 const DIAGRAM_MARGIN: i32 = 60;
@@ -380,213 +302,48 @@ fn spread(i: usize, count: usize, start: i32, end: i32) -> i32 {
     start + (i as i32) * (end - start) / (count as i32 - 1)
 }
 
-/// Draws a single panel — mesh, series and optional legend — onto its region.
-fn draw_panel(
-    area: &DrawingArea<BitMapBackend, Shift>,
-    panel: &Panel,
-    cfg: &ImageConfig,
-) -> Result<(), Box<dyn Error>> {
-    let (x_min, x_max) = panel.x_range;
-    let (y_min, y_max) = panel.y_range;
-    let mut chart = ChartBuilder::on(area)
-        .caption(&panel.title, (cfg.font_style, cfg.font_size).into_font())
-        .x_label_area_size(cfg.area_size)
-        .y_label_area_size(cfg.area_size)
-        .build_cartesian_2d(x_min..x_max, y_min..y_max)?;
-
-    chart.configure_mesh().draw()?;
-
-    for series in &panel.series {
-        draw_series(&mut chart, series)?;
-    }
-
-    if panel.show_legend {
-        chart
-            .configure_series_labels()
-            .background_style(WHITE.mix(0.8))
-            .border_style(BLACK)
-            .label_font((cfg.font_style, cfg.font_size))
-            .legend_area_size(cfg.area_size)
-            .position(SeriesLabelPosition::LowerRight)
-            .draw()?;
-    }
-
-    Ok(())
-}
-
-/// Draws one series, registering it for the legend when it carries a label.
-fn draw_series(
-    chart: &mut ChartContext<BitMapBackend, Cartesian2d<RangedCoordf32, RangedCoordf32>>,
-    series: &Series,
-) -> Result<(), Box<dyn Error>> {
-    match series {
-        Series::Line {
-            points,
-            color,
-            label,
-        } => {
-            let color = rgb(*color);
-            let annotation = chart.draw_series(LineSeries::new(points.iter().copied(), &color))?;
-            if let Some(label) = label {
-                annotation
-                    .label(label)
-                    .legend(move |(x, y)| Circle::new((x, y), 2, color.filled()));
-            }
-        }
-        Series::Points {
-            points,
-            color,
-            label,
-            radius,
-        } => {
-            let color = rgb(*color).filled();
-            let radius = *radius as i32;
-            let annotation =
-                chart.draw_series(points.iter().map(|&p| Circle::new(p, radius, color)))?;
-            if let Some(label) = label {
-                annotation
-                    .label(label)
-                    .legend(move |(x, y)| Circle::new((x, y), radius, color));
-            }
-        }
-    }
-
-    Ok(())
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::plot::scene::{Color, Figure, Panel, Series};
+    use crate::activations::RELU;
+    use crate::model::{NeuralNetwork, NeuronLayer};
+    use crate::plot::activations::DiagramOptions;
+    use ndarray::array;
 
     fn pixel_count(buffer: &[u8]) -> usize {
         buffer.len() / 3
     }
 
-    mod activation_diagram {
-        use super::*;
-        use crate::activations::RELU;
-        use crate::model::{NeuralNetwork, NeuronLayer};
-        use crate::plot::activations::DiagramOptions;
-        use ndarray::array;
-
-        fn diagram() -> crate::plot::activations::ActivationDiagram {
-            let net = NeuralNetwork {
-                layers: vec![NeuronLayer {
-                    weights: array![[1.0, 0.0], [-1.0, 0.0], [2.0, 0.0]],
-                    biases: array![0.0, 0.0, 0.0],
-                    activation: RELU.clone(),
-                }],
-            };
-            net.activation_diagram(array![1.0, 1.0].view(), &DiagramOptions::default())
-                .unwrap()
-        }
-
-        #[test]
-        fn to_image_draws_the_node_link_graph_onto_the_canvas() {
-            let image = diagram().to_image(&ImageConfig::new(300, 200)).unwrap();
-            assert_eq!((image.width, image.height), (300, 200));
-            assert_eq!(pixel_count(&image.bytes), 300 * 200);
-            // Nodes, edges and labels leave non-white pixels behind.
-            assert!(image.bytes.iter().any(|&byte| byte != 255));
-        }
-
-        #[test]
-        fn node_positions_lay_layers_left_to_right() {
-            let diagram = diagram();
-            let cfg = ImageConfig::new(300, 200);
-            let positions = diagram.node_positions(&cfg);
-            // Two columns (input, output), the output sitting right of the input.
-            assert_eq!(positions.len(), 2);
-            assert_eq!(positions[0].len(), 2);
-            assert_eq!(positions[1].len(), 3);
-            assert!(positions[0][0].0 < positions[1][0].0);
-        }
-    }
-
-    #[test]
-    fn to_image_draws_a_legend_for_labeled_line_and_point_series() {
-        // A labeled line and a labeled scatter, with the legend on: both legend
-        // markers (line and points) are rendered.
-        let figure = Figure::spatial(vec![Panel {
-            title: "Lines".to_string(),
-            x_range: (0.0, 10.0),
-            y_range: (0.0, 1.0),
-            show_legend: true,
-            series: vec![
-                Series::Line {
-                    points: vec![(0.0, 0.0), (10.0, 1.0)],
-                    color: Color::TRAIN,
-                    label: Some("Train".to_string()),
-                },
-                Series::Points {
-                    points: vec![(2.0, 0.3), (8.0, 0.7)],
-                    color: Color::category(0),
-                    label: Some("Class 0".to_string()),
-                    radius: 2,
-                },
-            ],
-        }]);
-
-        let cfg = ImageConfig::new(200, 150);
-        let image = figure.to_image(&cfg).unwrap();
-        assert_eq!((image.width, image.height), (200, 150));
-        assert_eq!(pixel_count(&image.bytes), 200 * 150);
-    }
-
-    #[test]
-    fn to_image_renders_points_and_unlabeled_series_without_legend() {
-        let figure = Figure::spatial(vec![Panel {
-            title: "Scatter".to_string(),
-            x_range: (0.0, 1.0),
-            y_range: (0.0, 1.0),
-            show_legend: false,
-            series: vec![
-                Series::Line {
-                    points: vec![(0.0, 0.0), (1.0, 1.0)],
-                    color: Color::TRAIN,
-                    label: None,
-                },
-                Series::Points {
-                    points: vec![(0.2, 0.2), (0.8, 0.8)],
-                    color: Color::category(0),
-                    label: Some("Class 0".to_string()),
-                    radius: 2,
-                },
-                Series::Points {
-                    points: vec![(0.5, 0.5)],
-                    color: Color::BOUNDARY,
-                    label: None,
-                    radius: 1,
-                },
-            ],
-        }]);
-
-        let cfg = ImageConfig::default();
-        let image = figure.to_image(&cfg).unwrap();
-        assert_eq!(pixel_count(&image.bytes), (cfg.width * cfg.height) as usize);
-    }
-
-    #[test]
-    fn to_image_stacks_multiple_panels() {
-        let panel = || Panel {
-            title: "Panel".to_string(),
-            x_range: (0.0, 1.0),
-            y_range: (0.0, 1.0),
-            show_legend: false,
-            series: Vec::new(),
+    fn diagram() -> ActivationDiagram {
+        let net = NeuralNetwork {
+            layers: vec![NeuronLayer {
+                weights: array![[1.0, 0.0], [-1.0, 0.0], [2.0, 0.0]],
+                biases: array![0.0, 0.0, 0.0],
+                activation: RELU.clone(),
+            }],
         };
-        let figure = Figure::spatial(vec![panel(), panel()]);
-
-        let image = figure.to_image(&ImageConfig::new(120, 120)).unwrap();
-        assert_eq!(pixel_count(&image.bytes), 120 * 120);
+        net.activation_diagram(array![1.0, 1.0].view(), &DiagramOptions::default())
+            .unwrap()
     }
 
     #[test]
-    fn to_image_of_an_empty_figure_is_blank() {
-        let figure = Figure::spatial(Vec::new());
-        let image = figure.to_image(&ImageConfig::new(64, 64)).unwrap();
-        // No panels drawn: every pixel stays white.
-        assert!(image.bytes.iter().all(|&byte| byte == 255));
+    fn to_image_draws_the_node_link_graph_onto_the_canvas() {
+        let image = diagram().to_image(&ImageConfig::new(300, 200)).unwrap();
+        assert_eq!((image.width, image.height), (300, 200));
+        assert_eq!(pixel_count(&image.bytes), 300 * 200);
+        // Nodes, edges and labels leave non-white pixels behind.
+        assert!(image.bytes.iter().any(|&byte| byte != 255));
+    }
+
+    #[test]
+    fn node_positions_lay_layers_left_to_right() {
+        let diagram = diagram();
+        let cfg = ImageConfig::new(300, 200);
+        let positions = diagram.node_positions(&cfg);
+        // Two columns (input, output), the output sitting right of the input.
+        assert_eq!(positions.len(), 2);
+        assert_eq!(positions[0].len(), 2);
+        assert_eq!(positions[1].len(), 3);
+        assert!(positions[0][0].0 < positions[1][0].0);
     }
 }
